+++ bin/runtime/src/lib.rs
@@ -6,6 +6,8 @@
 #[cfg(feature = "std")]
 include!(concat!(env!("OUT_DIR"), "/wasm_binary.rs"));
 
+mod chain_extension;
+
 use ark_bls12_381::Bls12_381;
 pub use frame_support::{
     construct_runtime, log, parameter_types,
@@ -20,7 +22,6 @@ pub use frame_support::{
     StorageValue,
 };
 use frame_support::{
-    log::error,
     sp_runtime::Perquintill,
     traits::{ConstU32, ConstU64, EqualPrivilegeOnly, SortedMembers, U128CurrencyToVote},
     weights::constants::WEIGHT_PER_MILLIS,
@@ -28,12 +29,7 @@ use frame_support::{
 };
 use frame_system::{EnsureRoot, EnsureSignedBy};
 pub use pallet_balances::Call as BalancesCall;
-use pallet_contracts::{
-    chain_extension::{
-        ChainExtension, Environment, Ext, InitState, RetVal, SysConfig, UncheckedFrom,
-    },
-    weights::WeightInfo,
-};
+use pallet_contracts::weights::WeightInfo;
 use pallet_contracts_primitives::{
     CodeUploadResult, ContractExecResult, ContractInstantiateResult, GetStorageResult,
 };
@@ -57,7 +53,7 @@ use sp_runtime::{
         OpaqueKeys, Verify,
     },
     transaction_validity::{TransactionSource, TransactionValidity},
-    ApplyExtrinsicResult, DispatchError, MultiSignature, RuntimeAppPublic,
+    ApplyExtrinsicResult, MultiSignature, RuntimeAppPublic,
 };
 pub use sp_runtime::{FixedPointNumber, Perbill, Permill};
 use sp_staking::EraIndex;
@@ -383,6 +379,9 @@ parameter_types! {
 }
 
 use sp_runtime::traits::Convert;
+
+use crate::chain_extension::AlephChainExtension;
+
 pub struct BalanceToU256;
 impl Convert<Balance, sp_core::U256> for BalanceToU256 {
     fn convert(balance: Balance) -> sp_core::U256 {
@@ -644,51 +643,6 @@ impl pallet_utility::Config for Runtime {
     type Call = Call;
     type WeightInfo = pallet_utility::weights::SubstrateWeight<Runtime>;
     type PalletsOrigin = OriginCaller;
-}
-
-const SNARCOS_CHAIN_EXT: u32 = 41;
-
-pub struct AlephChainExtension;
-impl ChainExtension<Runtime> for AlephChainExtension {
-    fn call<E: Ext>(func_id: u32, env: Environment<E, InitState>) -> Result<RetVal, DispatchError>
-    where
-        <E::T as SysConfig>::AccountId: UncheckedFrom<<E::T as SysConfig>::Hash> + AsRef<[u8]>,
-    {
-        match func_id {
-            SNARCOS_CHAIN_EXT => {
-                use pallet_snarcos::{Error, Pallet as Snarcos};
-
-                // The argument-passing-mode doesn't matter for now. All the data to runtime call
-                // are mocked now.
-                let mut env = env.buf_in_buf_out();
-                // After benchmarking is merged and `pallet_snarcos::WeightInfo` is available,
-                // use real weight here.
-                env.charge_weight(41)?;
-
-                match Snarcos::<Runtime>::bare_store_key([0u8; 4], [0u8; 8].to_vec()) {
-                    // In case `DispatchResultWithPostInfo` was returned (or some simpler
-                    // equivalent for `bare_store_key`), we could adjust weight. However, for the
-                    // storing key action it doesn't make sense.
-                    Ok(_) => {
-                        // Return status code for success.
-                        Ok(RetVal::Converging(0))
-                    }
-                    Err(Error::<Runtime>::VerificationKeyTooLong) => Ok(RetVal::Converging(1)),
-                    Err(Error::<Runtime>::IdentifierAlreadyInUse) => Ok(RetVal::Converging(2)),
-                    // Unknown error.
-                    _ => Ok(RetVal::Converging(3)),
-                }
-            }
-            _ => {
-                error!("Called an unregistered `func_id`: {}", func_id);
-                Err(DispatchError::Other("Unimplemented func_id"))
-            }
-        }
-    }
-
-    fn enabled() -> bool {
-        true
-    }
 }
 
 // Prints debug output of the `contracts` pallet to stdout if the node is started with `-lruntime::contracts=debug`.
